Began Implementation of Fold-Out Side Menu
Implementation begun. Sliding in/out works well, but side menu cannot
be clicked or interacted with. Not a problem with linking to the view
controller, since viewDidLoad runs. Seem to be problems with
NavigationViewController, DetailViews not rendering correctly. Might
switch back to original slide-out, and add swipe right gesture to open
and swipe left to close.