Share button added, DetailView contentSize fixed
notes:
DetailView: scrollview have a content size (CS) as well as a frame size
(FS), both of which are different. They scroll if CS is larger than FS.
If items are constrained to the width and height of the scrollview,
they are constrained to the FS. if items are constrained to the edges,
they are constrained to CS. Problem arose because the width of the FS
and the CS were not the same. Apple engineer helped by inserting a
structural view, with no height. View was constrained to the width of
the scrollview as well as the horizontal edges of it, thereby locking
the FS and the CS width as equal to each other. Magic.